https://github.com/jackdewinter/pymarkdown/issues/598 - refactoring LRD helper into links folder
