+++ publish/coverage.json
@@ -6,8 +6,8 @@
         "totalCovered": 3581
     },
     "lineLevel": {
-        "totalMeasured": 14451,
-        "totalCovered": 14451
+        "totalMeasured": 14479,
+        "totalCovered": 14479
     }
 }
